Unrar - Removed update if not required

+++ utils/unrar/unrar-installer.sh
@@ -18,17 +18,24 @@ URL=$(curl -s http://www.rarlab.com/rar_add.htm | grep -o '".*"' | grep 'unrarsr
 echo
 echo -e $YELLOW"URL Found: ${URL}"$ENDCOLOR
 VERSION=$(echo $URL | grep -oP '(?<=unrarsrc-).*(?=.tar)')
-echo -e $YELLOW"Version Found: $VERSION"$ENDCOLOR
-echo
-sudo curl "${URL}" | tar -xz
-cd unrar
-echo
-sudo apt-get remove unrar -y
-echo 'Starting Build'
-sudo make
-sudo checkinstall -y --pkgversion="1:${VERSION}" --pakdir='/var/cache/apt/archives'
+echo -e $YELLOW"Downloaded Version: 1:${VERSION}-1"$ENDCOLOR
+CURRENTUNRAR=$(dpkg -s unrar | grep Version | cut -d " " -f2-)
+echo -e $YELLOW"Installed Version:  $CURRENTUNRAR"$ENDCOLOR
 
-rm -r $(pwd)
-cd $SCRIPTPATH
-source $SCRIPTPATH/inc/app-constant-reset.sh
-echo -e $GREEN"---> Unrar installation complete."$ENDCOLOR
+if [[ ! $CURRENTUNRAR = "1:${VERSION}-1" ]] ; then
+    echo
+    sudo curl "${URL}" | tar -xz
+    cd unrar
+    echo
+    sudo apt-get remove unrar -y
+    echo 'Starting Build'
+    sudo make
+    sudo checkinstall -y --pkgversion="1:${VERSION}" --pakdir='/var/cache/apt/archives'
+
+    rm -r $(pwd)
+    cd $SCRIPTPATH
+    source $SCRIPTPATH/inc/app-constant-reset.sh
+    echo -e $GREEN"---> Unrar installation complete."$ENDCOLOR
+else
+    echo -e $GREEN"---> Unrar update not required."$ENDCOLOR
+fi
